+++ src/main/java/lagatrix/server/connection/communicators/AESCommunicator.java
@@ -13,17 +13,17 @@ import javax.crypto.IllegalBlockSizeException;
 import javax.crypto.NoSuchPaddingException;
 import javax.crypto.SealedObject;
 import javax.crypto.SecretKey;
+import lagatrix.server.entities.actions.ActionsEnum;
 import lagatrix.server.entities.connection.Request;
 import lagatrix.server.entities.connection.Response;
+import lagatrix.server.exceptions.connection.AlgorithmException;
 import lagatrix.server.exceptions.connection.BadClassFormatException;
 import lagatrix.server.exceptions.connection.ConnectionInOutException;
-import lagatrix.server.exceptions.connection.ObtainRequestException;
-import lagatrix.server.exceptions.connection.SendResponseException;
 
 /**
- * This class manage the Rquest and Response objects with RSA encryption.
+ * This class manage the Rquest and Response objects with AES encryption.
  *
- * @author javier
+ * @author javierfh03
  * @since 0.2
  */
 public class AESCommunicator implements CommunicatorBase {
@@ -49,13 +49,12 @@ public class AESCommunicator implements CommunicatorBase {
             this.in = new ObjectInputStream(socket.getInputStream());
             this.out = new ObjectOutputStream(socket.getOutputStream());
         } catch (IOException ex) {
-            throw new ConnectionInOutException(ConnectionInOutException.getMessage(this.getClass(), 
-                    "AES problem I/O when open connection", socket.getInetAddress().toString()));
+            throw new ConnectionInOutException(ConnectionInOutException.getMessageIO("AES", ActionsEnum.OPEN));
         }
     }
 
     @Override
-    public Request obtainRequest() throws ObtainRequestException, BadClassFormatException {
+    public Request obtainRequest() throws ConnectionInOutException, AlgorithmException, BadClassFormatException {
         Cipher encryptCipher;
         SealedObject sealedObject;
         try {
@@ -66,16 +65,17 @@ public class AESCommunicator implements CommunicatorBase {
             sealedObject = (SealedObject) in.readObject();
         
             return (Request) sealedObject.getObject(encryptCipher);
-        } catch (NoSuchAlgorithmException | InvalidKeyException | NoSuchPaddingException | IOException | IllegalBlockSizeException ex) {
-            throw new ObtainRequestException("AES");
+        } catch (NoSuchAlgorithmException | InvalidKeyException | NoSuchPaddingException | IllegalBlockSizeException ex) {
+            throw new AlgorithmException(AlgorithmException.getMessageAlgorithm(ex, "AES"));
         } catch (ClassNotFoundException | BadPaddingException ex) {
-            throw new BadClassFormatException(BadClassFormatException.getMessage(this.getClass(), 
-                    "unexpected class", socket.getInetAddress().toString()));
+            throw new BadClassFormatException("unexpected class");
+        } catch (IOException ex) {
+            throw new ConnectionInOutException(ConnectionInOutException.getMessageIO("AES", ActionsEnum.RECEIVE));
         }
     }
 
     @Override
-    public void sendResponse(Response response) throws SendResponseException {
+    public void sendResponse(Response response) throws ConnectionInOutException, AlgorithmException {
         Cipher encryptCipher;
         SealedObject sealedObject;
         
@@ -86,9 +86,11 @@ public class AESCommunicator implements CommunicatorBase {
             sealedObject = new SealedObject( (Serializable) response, encryptCipher);
             
             out.writeObject(sealedObject);
-        } catch (NoSuchAlgorithmException | InvalidKeyException | NoSuchPaddingException | IOException | IllegalBlockSizeException ex) {
-            throw new SendResponseException("AES");
-        }  
+        } catch (NoSuchAlgorithmException | InvalidKeyException | NoSuchPaddingException | IllegalBlockSizeException ex) {
+            throw new AlgorithmException("AES");
+        }  catch (IOException ex) {
+            throw new ConnectionInOutException(ConnectionInOutException.getMessageIO("AES", ActionsEnum.SEND));
+        }
     }
 
     @Override
@@ -97,8 +99,7 @@ public class AESCommunicator implements CommunicatorBase {
             out.close();
             in.close();
         } catch (IOException ex) {
-            throw new ConnectionInOutException(ConnectionInOutException.getMessage(this.getClass(), 
-                    "AES problem I/O when close connection", socket.getInetAddress().toString()));
+            throw new ConnectionInOutException(ConnectionInOutException.getMessageIO("AES", ActionsEnum.CLOSE));
         }
     }
 
@@ -108,8 +109,7 @@ public class AESCommunicator implements CommunicatorBase {
             close();
             socket.close();
         } catch (IOException ex) {
-            throw new ConnectionInOutException(ConnectionInOutException.getMessage(this.getClass(), 
-                    "AES problem I/O when close the client connection", socket.getInetAddress().toString()));
+            throw new ConnectionInOutException(ConnectionInOutException.getMessageIO("AES", ActionsEnum.CLOSE));
         }
     }