+++ src/main/java/lagatrix/server/exceptions/connection/AlgorithmException.java
@@ -12,4 +12,15 @@ public class AlgorithmException extends ConnectionException {
         super(description);
     }
 
+    /**
+     * This method obtain the error message.
+     * 
+     * @param excpetion The exception of cypher who excepted.
+     * @param algorithm The algorithm who used.
+     * @return The message.
+     */
+    public static String getMessageAlgorithm(Exception excpetion, String algorithm) {
+        return String.format("%s have an this error when cypher: %s", algorithm, 
+                excpetion.getClass().getSimpleName());
+    }
 }